refactor: melhorando nome dos componentes novos
